Implement Swarm node manager
This implements the Swarm node manager, which is responsible for starting and
closing the Swarm process properly. I'm trying to follow as closely as possible
the style I found in other files, although I'm a little concerned with some
patterns such as the use of global state and singletons. Those must be debated
later. There are still some things to implement and test in order for this to
be considered done.